update
- added chams
- reworked color class
- added material
- added materialsystem
- added new entity shit
- added variable namespace

+++ utilities/render/render.cpp
@@ -9,7 +9,7 @@ bool Render::InitRenderer() {
 
 void Render::DrawFilledRect(int x, int y, int w, int h, Color _col) {
 	D3DRECT rect = { x,y, x + w, y + h };
-	g_DirectX.pDevice->Clear(1, &rect, D3DCLEAR_TARGET, D3DCOLOR_ARGB(_col.a, _col.r, _col.g, _col.b), 0.f, 0.f);
+	g_DirectX.pDevice->Clear(1, &rect, D3DCLEAR_TARGET, D3DCOLOR_ARGB((int)_col.a, (int)_col.r, (int)_col.g, (int)_col.b), 0.f, 0.f);
 }
 
 void Render::DrawLine(int x1, int y1, int x2, int y2, int width, Color _col) {
@@ -21,7 +21,7 @@ void Render::DrawLine(int x1, int y1, int x2, int y2, int width, Color _col) {
 	Line[1] = D3DXVECTOR2(x2, y2);
 
 	g_Render.g_pLine->SetWidth(width);
-	g_Render.g_pLine->Draw(Line, 2, D3DCOLOR_ARGB(_col.a, _col.r, _col.g, _col.b));
+	g_Render.g_pLine->Draw(Line, 2, D3DCOLOR_ARGB((int)_col.a, (int)_col.r, (int)_col.g, (int)_col.b));
 }
 
 void Render::DrawLine(Vec2D src, Vec2D dst, int width, Color _col) {
@@ -42,10 +42,10 @@ void Render::Text(LPD3DXFONT font, const char* text, float x, float y, int scale
 	RECT rect;
 
 	SetRect(&rect, x + 1, y + 1, x + 1, y + 1);
-	g_Render.pEspFont->DrawTextA(NULL, text, -1, &rect, DT_CENTER | DT_NOCLIP, D3DCOLOR_ARGB(_col.a, 0, 0, 0));
+	g_Render.pEspFont->DrawTextA(NULL, text, -1, &rect, DT_CENTER | DT_NOCLIP, D3DCOLOR_ARGB((int)_col.a, 0, 0, 0));
 
 	SetRect(&rect, x, y, x, y);
-	g_Render.pEspFont->DrawTextA(NULL, text, -1, &rect, DT_CENTER | DT_NOCLIP, D3DCOLOR_ARGB(_col.a, _col.r, _col.g, _col.b));
+	g_Render.pEspFont->DrawTextA(NULL, text, -1, &rect, DT_CENTER | DT_NOCLIP, D3DCOLOR_ARGB((int)_col.a, (int)_col.r, (int)_col.g, (int)_col.b));
 }
 
 Vec2D Render::TextSize(LPD3DXFONT font, const char* text) {
